stock_portfolio_analyzer/requirements.txt
pandas
yfinance
scipy
streamlit
requests
beautifulsoup4
lxml
newsapi-python
python-dotenv
